Use one risky card for both pool positions page and investor proposals page
