--- src/components/cards/proposal/RiskyInvestor/RiskyProposalInvestorCard.stories.tsx
@@ -1,5 +1,0 @@
-/* eslint-disable */
-import { storiesOf } from '@storybook/react';
-import RiskyProposalInvestorCard from './index';
-
-// storiesOf('RiskyProposalInvestorCard', module).add('default', () => <RiskyProposalInvestorCard />);
