+++ src/pages/InvestmentRiskyProposals/ProposalsList.tsx
@@ -1,18 +1,65 @@
-import { FC } from "react"
+import { FC, useMemo } from "react"
 import { PulseSpinner } from "react-spinners-kit"
 
+import { useActiveWeb3React } from "hooks"
+import { usePoolContract } from "hooks/usePool"
+import { RiskyProposal } from "constants/interfaces_v2"
+import { useRiskyProposalContract } from "hooks/useContract"
 import useInvestorRiskyProposals from "hooks/useInvestorRiskyProposals"
-import RiskyProposalInvestorCard from "components/cards/proposal/RiskyInvestor"
+
+import RiskyProposalCard from "components/cards/proposal/Risky"
+
 import S from "./styled"
 
 interface IProps {
   activePools: string[]
 }
 
+interface IRiskyCardInializer {
+  index: number
+  account: string
+  poolAddress: string
+  proposal: RiskyProposal
+}
+
+const RiskyProposalCardInializer: FC<IRiskyCardInializer> = ({
+  index,
+  account,
+  poolAddress,
+  proposal,
+}) => {
+  const [proposalPool] = useRiskyProposalContract(poolAddress)
+  const [, poolInfo] = usePoolContract(poolAddress)
+
+  const isTrader = useMemo<boolean>(() => {
+    if (!account || !poolInfo) {
+      return false
+    }
+
+    return account === poolInfo.parameters.trader
+  }, [account, poolInfo])
+
+  if (!proposal || !poolInfo || !proposalPool) {
+    return null
+  }
+
+  return (
+    <RiskyProposalCard
+      proposalId={index}
+      poolInfo={poolInfo}
+      isTrader={isTrader}
+      proposal={proposal}
+      poolAddress={poolAddress}
+      proposalPool={proposalPool}
+    />
+  )
+}
+
 const InvestmentRiskyProposalsList: FC<IProps> = ({ activePools }) => {
+  const { account } = useActiveWeb3React()
   const [proposals, fetched] = useInvestorRiskyProposals(activePools)
 
-  if (!fetched) {
+  if (!fetched || !account) {
     return (
       <S.Content>
         <PulseSpinner />
@@ -32,10 +79,11 @@ const InvestmentRiskyProposalsList: FC<IProps> = ({ activePools }) => {
     <>
       <S.List>
         {proposals.map((p, i) => (
-          <RiskyProposalInvestorCard
+          <RiskyProposalCardInializer
             key={p.poolAddress + i}
+            index={i + 1}
+            account={account}
             proposal={p.proposal}
-            proposalId={i}
             poolAddress={p.poolAddress}
           />
         ))}
